fix: 修存档牌表指纹读写不一致
牌表指纹是 uint64，但存档 JSON 数值只有 int64 精确域：写出侧原先直接输出
unsigned 十进制，指纹 ≥2^63 时被 pjh_json 退化为 double，读取端 try_as_int
失败报 StructureError，约 50% 概率的自定义牌表存档写出后读不回。

写出侧改按 int64 位型（static_cast）承载，高位指纹落成负十进制，读取端既有
static_cast<uint64_t> 逐位还原；不改哈希算法。相比十进制字符串方案改动更小，
且新旧二进制双向兼容（新 reader 可读旧档、旧 reader 可读新档），无 double
往返与精度丢失。

+++ src/save/writer.hpp
@@ -243,8 +243,11 @@ namespace tkw
 
             std::ostringstream os;
             os << "{\"format\":" << jstr(kFormat) << ",\"version\":" << kVersion;
+            // 指纹按 int64 位型写出：JSON 数值只有 int64 精确域，高位指纹
+            // （≥2^63）须落成负十进制，读取端再逐位还原为 uint64。
             os << ",\"deck\":{\"name\":" << jstr(deck_name)
-               << ",\"hash\":" << deck_hash(g.catalog) << "}";
+               << ",\"hash\":" << static_cast<std::int64_t>(deck_hash(g.catalog))
+               << "}";
             os << ",\"rules\":{"
                << "\"draw_per_turn\":" << r.draw_per_turn
                << ",\"sha_limit\":" << r.sha_limit
